+++ backend/src/main/java/com/youmorry/expensetracker/infrastructure/persistence/JdbcConfiguration.java
@@ -1,5 +1,6 @@
 package com.youmorry.expensetracker.infrastructure.persistence;
 
+import com.youmorry.expensetracker.infrastructure.persistence.converter.IdConverters;
 import java.util.List;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jdbc.repository.config.AbstractJdbcConfiguration;
